size_t -> unsigned int

+++ include/Error_handler.hpp
@@ -8,12 +8,12 @@ namespace Error
 {
     struct ErrorInfo
     {
-        const size_t line;
+        const unsigned int line;
         const std::string where;
         const std::string message;
     };
     void report() noexcept;
-    void add(const size_t line, const std::string where, const std::string message) noexcept;
+    void add(const unsigned int line, const std::string where, const std::string message) noexcept;
     void clear() noexcept;
     extern bool hadError;
     extern bool hadRuntimeError;
