+++ src/Error_handler.cpp
@@ -5,7 +5,7 @@ namespace Error
     std::vector<ErrorInfo> exceptionList{};
     bool hadError = false;
     bool hadRuntimeError = false;
-    void add(const size_t line, const std::string where, const std::string message) noexcept
+    void add(const unsigned int line, const std::string where, const std::string message) noexcept
     {
         exceptionList.push_back({line, where, message});
         hadError = true;
